- added kamikaze behaviour - added static defence Behaviour - masses of whitespace added to make code prettier - macros moved into macro.h - enums moved into enums.h - added macro definition to prevent VS2005 spamming warnings about security and suggesting alternatives that linux cant use - fixed a slew of validunitID checks - got rid of tonnes of redundant or pointless code - removed unnecesary and duplicate headers in include.h - added lots of comments to headers and include directives.
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@4893 37977431-3df6-0310-b722-df95706aa16b

+++ AI/Global/NTai/AI/NTai/Tasks/CKeywordConstructionTask.cpp
@@ -245,7 +245,7 @@ void CKeywordConstructionTask::Build(){
 
 		int* funits = new int[5000];
 
-		int fnum = G->cb->GetFriendlyUnits(funits,unitpos,building->GetExclusionRange());
+		int fnum = G->cb->GetFriendlyUnits(funits,unitpos,(float)building->GetExclusionRange());
 
 		if(fnum > 1){
 			//
